requirements.txt
torch
torchvision
torchcam
Pillow
streamlit
opencv-python-headless
numpy
